Renamed use-smart-transaction to use-transactions
