+++ src/app/(pages)/profile/claim-winning/_components/claimable-prizes.tsx
@@ -11,8 +11,7 @@ import SectionContent from './section-content'
 import SectionTitle from './section-title'
 import { useClaimablePools } from './use-claimable-pools'
 import { useAppStore } from '@/app/_client/providers/app-store.provider'
-import { getConfig } from '@/app/_client/providers/configs/wagmi.config'
-import useTransactions from '@/app/_client/hooks/use-smart-transaction'
+import useTransactions from '@/app/_client/hooks/use-transactions'
 import { currentPoolAddress } from '@/app/_server/blockchain/server-config'
 import { poolAbi } from '@/types/contracts'
