+++ src/app/@bottombar/_components/registered-dropdown.list.tsx
@@ -16,8 +16,7 @@ import { getAbiItem } from 'viem'
 import RegisteredDropdownItem from './registered-dropdown.item'
 import type { RegisteredDropdownItemConfig } from './registered-dropdown.list.config'
 import { dropdownItemsConfig } from './registered-dropdown.list.config'
-import { getConfig } from '@/app/_client/providers/configs/wagmi.config'
-import useTransactions from '@/app/_client/hooks/use-smart-transaction'
+import useTransactions from '@/app/_client/hooks/use-transactions'
 import { poolAbi } from '@/types/contracts'
 import { currentPoolAddress } from '@/app/_server/blockchain/server-config'
